Update Tool-Solutions/docker for 22.04 release
Squashed commit:
* Removes tf_cnn_benchmarks from TF build.
* Uses minimal welcome message for early build stages.
* Adds changelog for tensorflow-aarch64 and pytorch-aarch64
* Updates CHANGELOG.md with unreleased changes ready to upstream.
* Updates builds to use oneDNN 2.6-rc.
* Add capability to TF examples to load model in SavedModel format
* Updates oneDNN to v2.6 release.
* Downgrade torchvision to 0.11.1
* Use non-constant weights in ACL FC when training
* Build TF with threadpool support in oneDNN and ACL
* Do not download model if it already exists

+++ docker/pytorch-aarch64/examples/executor/model.py
@@ -1,5 +1,5 @@
 # *******************************************************************************
-# Copyright 2021 Arm Limited and affiliates.
+# Copyright 2021-2022 Arm Limited and affiliates.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -94,15 +94,19 @@ class Model:
         model_url = model_descriptor["model"][0]["source"]
         model_name = model_descriptor["model"][0]["name"]
 
-        try:
-            # Download the model
-            urllib.request.urlretrieve(
-                model_url,
-                model_name,
-                DownloadProgressBar("Downloading: " + model_name + "..."),
-            )
-        except:  # pylint: disable=bare-except
-            return False
+        # check to see whether the model exists in the
+        # current directory and if it exists then we do
+        # not need to download it
+        if not os.path.isfile(model_name):
+            try:
+                # Download the model
+                urllib.request.urlretrieve(
+                    model_url,
+                    model_name,
+                    DownloadProgressBar("Downloading: " + model_name + "..."),
+                )
+            except:  # pylint: disable=bare-except
+                return False
 
         if "class" in model_descriptor["model"][0]:
             model_class = model_descriptor["model"][0]["class"]
